Make `run_fibonacci()` an in-process test
Previously it was a bit hard to debug, since we couldn't see the child process' stdout when running the test (unless we switch it to `inherit()` temporarily).

It means we can't reuse some code from the bin, so I extracted the most nontrivial part into a shared function `ProgramROM::from_file`.